Pin create_input to version of reference solution
Current 'latest' dumps differ from the dump version used to generate the
reference solution small-ref.bz2.  New dumps are created around the 1st
and 20th of each month.

Note - the dumps that begin on the 1st of the month are considered a full
dump while those that begin on the 20th of the month are considered a
parial dump. At least that's what the documentation says.  I haven't
noticed a difference.  https://meta.wikimedia.org/wiki/Data_dumps/Dump_frequency
The most recent parital and full dump are both around the same size (10mb
difference in size), and appear to be the same format.  Still you may want
to specifify which version of the data the contest will be using, even if
that version is a dump which doesn't exist yet so precomputing the result
doesn't work.

Also set urls to https.  Wikimedia has hsts policy so it will be
negotiated anyway.